Move PubSubTransport back again
